+++ src/backend/commands/player/toxic/kick_from_vehicle.cpp
@@ -22,7 +22,7 @@ namespace big
 			{
 				// vehicle hasn't synced yet, use TSE
 				const size_t arg_count  = 10;
-				int64_t args[arg_count] = {(int64_t)eRemoteEvent::VehicleKick, self::id, 1 << player->id(), 0, 0, 0, 0, 0, 0, 0};
+				int64_t args[arg_count] = {(int64_t)eRemoteEvent::VehicleKick, (int64_t)self::id, 1i64 << player->id(), 0, 0, 0, 0, 0, 0, 0};
 
 				g_pointers->m_gta.m_trigger_script_event(1, args, arg_count, 1 << player->id(), (int)eRemoteEvent::VehicleKick);
 			}
