+++ src/util/globals.hpp
@@ -13,7 +13,7 @@ namespace big::globals
 	{
 		constexpr size_t arg_count = 4;
 		int64_t args[arg_count]    = {static_cast<int64_t>(eRemoteEvent::ClearWantedLevel),
-		       self::id, 1 << target,
+		       (int64_t)self::id, 1i64 << target,
 		       (int64_t)scr_globals::gpbd_fm_3.as<GPBD_FM_3*>()->Entries[target].ScriptEventReplayProtectionCounter};
 
 		g_pointers->m_gta.m_trigger_script_event(1, args, arg_count, 1 << target, (int)eRemoteEvent::ClearWantedLevel);
@@ -24,7 +24,7 @@ namespace big::globals
 		constexpr size_t arg_count = 8;
 		int64_t args[arg_count]    = {static_cast<int64_t>(eRemoteEvent::RemoteOffradar),
 		       (int64_t)self::id,
-		       1 << target,
+		       1i64 << target,
 		       (int64_t)(NETWORK::GET_NETWORK_TIME() + 1),
 		       0,
 		       true,
